# api-document.yaml
openapi: 3.0.2
info:
  title: CA Tech Dojo API
  description: |-
    <b>CA Tech Dojo API仕様</b><br>
    まずはこのAPI仕様に沿って機能を実装してみましょう。<br><br>
    なお、実装に際してランキング機能の実装が必要となります。<br>
    ランキングの実装にMySQLではなくredisを利用することも可能です。<br>
    MySQLのORDER BYを利用するか、redisのZSETを利用して実装しましょう。
  version: 1.0.0
servers:
  - url: http://localhost:8080/
...
paths:
...
  /user/get:
    get:
      tags:
        - user
      summary: ユーザ情報取得API
      description: |
        ユーザ情報を取得します。
        「ユーザの認証と特定」の処理はリクエストヘッダの`x-token`を読み取ってデータベースに照会をします。
      parameters:
        - name: x-token
          in: header
          description: 認証トークン
          required: true
          schema:
            type: string
      responses:
        200:
          description: A successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGetResponse'
...
components:
  schemas:
...
    UserGetResponse:
      type: object
      properties:
        id:
          type: string
          description: ユーザID
        name:
          type: string
          description: ユーザ名
        highScore:
          type: integer
          description: ハイスコア
        coin:
          type: integer
          description: 所持コイン
...